Add likers list

+++ src/main/java/com/inst/service/impl/ImageServiceImpl.java
@@ -1,8 +1,9 @@
 package com.inst.service.impl;
 
 import com.inst.entity.Image;
-import com.inst.entity.Like;
+import com.inst.entity.Luke;
 import com.inst.entity.User;
+import com.inst.exception.EntityNotFoundException;
 import com.inst.repository.ImageRepository;
 import com.inst.repository.LikeRepository;
 import com.inst.service.ImageService;
@@ -10,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Collection;
+import java.util.List;
 
 @Service
 public class ImageServiceImpl implements ImageService {
@@ -27,6 +29,8 @@ public class ImageServiceImpl implements ImageService {
 
     @Override
     public void create(Image image) {
+        if (image == null)
+            throw new EntityNotFoundException("Image entity is null");
         imageRepository.create(image);
     }
 
@@ -42,7 +46,7 @@ public class ImageServiceImpl implements ImageService {
 
     @Override
     public boolean containLikeByUser(User user, Image image) {
-        for (Like like : image.getLikes()) {
+        for (Luke like : image.getLikes()) {
             if(like.getUser().getId() == user.getId())
                 return true;
         }
@@ -50,17 +54,33 @@ public class ImageServiceImpl implements ImageService {
     }
 
     @Override
-    public Image findById(int i) {
-        return imageRepository.findById(i);
+    public Image findById(int id) {
+        if( id < 0)
+            throw new IllegalArgumentException("Image id is less than 0");
+        return imageRepository.findById(id);
     }
 
     @Override
-    public void likeImage(Like like) {
+    public void likeImage(Luke like) {
+        if (like == null)
+            throw new EntityNotFoundException("Like entity is null");
         likeRepository.create(like);
     }
 
     @Override
-    public void unlikeImage(Like like) {
+    public void unlikeImage(Luke like) {
+        if (like == null)
+            throw new EntityNotFoundException("Like entity is null");
         likeRepository.delete(like);
+    }
+
+    @Override
+    public List<User> findAllImageLikers(int imageId) {
+        if( imageId < 0)
+            throw new IllegalArgumentException("Image id is less than 0");
+
+        Image image = this.findById(imageId);
+
+        return likeRepository.findAllImageLikers(image);
     }
 }
